requirements.txt
streamlit==1.28.1
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3